requirements.txt
solana==0.36.9
solders==0.26.0
requests==2.32.3
python-dotenv==1.0.1
websockets==12.0
cryptography==43.0.1
